useeffect and header refactor; add mui class to signin page
