+++ packages/client/package.json
@@ -47,6 +47,7 @@
     "@typescript-eslint/parser": "^5.35.1",
     "@vitejs/plugin-react": "^2.0.1",
     "eslint": "^8.23.0",
+    "eslint-plugin-react-hooks": "^5.2.0",
     "jest": "^28",
     "jest-environment-jsdom": "^29.0.1",
     "lefthook": "^1.3.9",
