+++ packages/client/src/pages/ForumPage.tsx
@@ -3,7 +3,7 @@ import { Helmet } from 'react-helmet'
 import { Link } from 'react-router-dom'
 import { usePage } from '../hooks/usePage'
 import { PageInitArgs } from '../routes'
-import { Header } from '../components/Header'
+import { Header } from '../components/Header/Header'
 import {
   Container,
   Typography,
