+++ packages/client/src/pages/LeaderboardPage.tsx
@@ -1,6 +1,6 @@
 import { Helmet } from 'react-helmet'
 import { PageInitArgs } from '../routes'
-import { Header } from '../components/Header'
+import { Header } from '../components/Header/Header'
 import {
   Container,
   Typography,
