+++ packages/client/src/pages/ProfilePage.tsx
@@ -6,7 +6,7 @@ import { ChangePasswordForm } from '../components/ChangePasswordForm/ChangePassw
 import { Typography } from '@material-ui/core'
 import { ChangeProfileForm } from '../components/ProfileForm/ProfileForm'
 import { AvatarInput } from '../components/AvatarInput/AvatarInput'
-import { Header } from '../components/Header'
+import { Header } from '../components/Header/Header'
 
 export const ProfilePage = () => {
   return (
