+++ packages/server/package.json
@@ -28,6 +28,7 @@
     "@typescript-eslint/parser": "^5.35.1",
     "babel-jest": "^29.0.1",
     "eslint": "^8.23.0",
+    "eslint-plugin-react-hooks": "^5.2.0",
     "jest": "^28",
     "nodemon": "^2.0.19",
     "prettier": "^2.7.1",
